requirements.txt
Babel==2.2.0
PyYAML==3.11
argparse==1.2.1
cliff==1.15.0
cmd2==0.6.8
configparser
debtcollector==1.2.0
funcsigs==0.4
iso8601==0.1.11
keystoneauth1==2.2.0
monotonic==0.5
msgpack-python==0.4.6
netaddr==0.7.18
netifaces==0.10.4
oslo.config==3.2.1
oslo.i18n==1.5.0
oslo.serialization==1.4.0
oslo.utils==1.4.0
pbr==0.11.1
prettytable==0.7.2
pyparsing==2.0.7
python-keystoneclient==2.1.1
python-neutronclient==2.3.11
pytz==2015.7
requests==2.9.1
simplejson==3.8.1
six==1.10.0
stevedore==1.10.0
unicodecsv==0.14.1
wrapt==1.10.6
wsgiref==0.1.2

